refactor: adjusts folder and file organization
