+++ src/components/betsCard/index.tsx
@@ -1,21 +1,22 @@
-import { IProduct } from '@/pages/types/product';
 import Image from 'next/image';
+import { IProduct } from '@/types/product';
 
-interface ISandalCardProps {
+interface IBetsCardProps {
   products: IProduct[];
 }
 
-export function SandalCard({ products }: ISandalCardProps) {
+export function BetsCard({ products }: IBetsCardProps) {
+  const PRODUCT_ITEMS_LIMIT = 5;
   return (
-    <div className="grid grid-cols-4 w-full gap-1">
-      {products.map((product, index) => (
-        <div key={index} className="flex flex-col w-full">
-          <div className="relative w-full aspect-square">
+    <div className="flex justify-between w-full overflow-x-auto gap-1">
+      {products.slice(0, PRODUCT_ITEMS_LIMIT).map((product, index) => (
+        <div key={index} className="flex flex-col min-w-72 w-full">
+          <div className="relative w-full mb-2 h-96">
             <Image
               src={product.images[0].url}
               alt={product.name}
               layout="fill"
-              className="object-cover"
+              className="w-full h-full object-cover"
             />
           </div>
           <div className="flex text-xs px-3 py-3 w-full">
